backend/requirements.txt
# Flask and web server
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Image processing
Pillow==10.0.1
pytesseract==0.3.10
python-magic==0.4.27

# Google Gemini API
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0